Move `ExtendedNode` to `NodeMethods` for consistency.
