+++ semanticcpg/src/main/scala/io/shiftleft/semanticcpg/language/package.scala
@@ -5,7 +5,7 @@ import io.shiftleft.codepropertygraph.Cpg
 import io.shiftleft.codepropertygraph.generated.nodes
 import io.shiftleft.codepropertygraph.generated.nodes.{Node, StoredNode}
 import io.shiftleft.semanticcpg.language.callgraphextension.{Call, Method}
-import io.shiftleft.semanticcpg.language.nodemethods.{AstNodeMethods, WithinMethodMethods}
+import io.shiftleft.semanticcpg.language.nodemethods.{AstNodeMethods, NodeMethods, WithinMethodMethods}
 import io.shiftleft.semanticcpg.language.types.structure._
 import io.shiftleft.semanticcpg.language.types.expressions._
 import io.shiftleft.semanticcpg.language.types.expressions.generalizations._
@@ -22,6 +22,8 @@ package object language {
 
   // Implicit conversions from generated node types. We use these to add methods
   // to generated node types.
+
+  implicit def toExtendedNode(node: Node): NodeMethods = new NodeMethods(node)
 
   implicit def withMethodMethodsQp(node: nodes.WithinMethod): WithinMethodMethods =
     new WithinMethodMethods(node)
@@ -186,23 +188,5 @@ package object language {
 
   implicit def toCallForCallGraph(steps: Steps[nodes.Call]): Call =
     new Call(steps.raw)
-
-  // Locations
-
-  implicit def toExtendedNode(node: Node): ExtendedNode =
-    new ExtendedNode(node)
-
-  case class ExtendedNode(node: Node) {
-
-    def location: nodes.NewLocation = {
-      node match {
-        case storedNode: StoredNode =>
-          LocationCreator(storedNode)
-        case _ =>
-          LocationCreator.emptyLocation("", None)
-
-      }
-    }
-  }
 
 }
